\name{SegRegPar-package}
\alias{SegRegPar-package}
\alias{SegRegPar}
\docType{package}
\title{
\packageTitle{SegRegPar}
}
\description{
\packageDescription{SegRegPar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SegRegPar}
\packageIndices{SegRegPar}
}
\author{
\packageAuthor{SegRegPar}

Maintainer: \packageMaintainer{SegRegPar}
}
\references{
}
	\keyword{ package }
\seealso{
}
\examples{
}
